Don't depend on servlet classes
This seems strange at first, but it is so that the servlet plugin can
be loaded on the root class loader along with informant jar.  Other
options include (a) add servlet classes to the root class loader as
well, (b) add servlet plugin into the application war file, (c) app
server specific solutions, e.g. one for Tomcat, one for WebLogic, etc.
(a), (b) and (c) all cause class loader nightmare, so even though this
solution makes the code a bit uglier, it is a major win to avoid class
loader nightmare.