Add caching of third-party versioned web resources
This takes care of the majority of web resources.  At some point should
decide on an approach for safely caching non-third-party web resources
that are versioned as part of informant.  One option is to add the
informant version to the url, but need to do this in a way that I won't
forget to increment the version in the resource urls with each new
version of informant.