Write large text to rolling file instead of db
This makes it much easier and more efficient to prevent uncontrolled
data growth.  Still need to rollover data in db, but this data is much
smaller and now the large text data won't push out the smaller summary
data.  The summary data in the db can still be searched after the detail
data has been rolled over, in which case the user will see the summary
data, but will see a message that the detail data has been rolled over.