Active trace details are not being displayed
If an active trace is selected on the chart and the active trace
completes before hitting the 'details' button then non-compliant
json is returned from the server.

Also change Trace 'id' column to be a primary key and merge completed
trace records on top of stuck trace records. Also change terminology
"live trace" to "active trace".