Fix PermGen OutOfMemoryError in unit tests
There are three places that are causing retention of the
IsolatedWeavingClassLoader.  Since each test class runs in its own
IsolatedWeavingClassLoader, these build up quickly and cause PermGen
OutOfMemoryError.

One of these places causing retention is the ThreadLocal holding the
spanStack, which isn't needed anymore anyways.

Another place is java.sql.DriverManager which retains all registered
Drivers.  To fix this, the H2 driver needs to be deregistered when it is
closed.

The last place causing retention is in the AspectJ library, see
https://bugs.eclipse.org/bugs/show_bug.cgi?id=373195.  For now this has
to be hacked around as explained in the bugzilla record.