Change agent argument db.file to data.dir
This is in preparation for aspectj cache and it will be easier to
specify a single directory instead of multiple file locations.

Also, set the default data.dir to the directory that informant.jar
is in, instead of the process 'current dir' (which is the annoying
C:\Windows\system32 when running tomcat as a service).