Add informant 'packages' to simplify install
Add underlying support for developers to be able to create informant
'packages' (e.g. informant-for-web, informant-for-my-app) to simplify
distribution and installation. These packages should be a single jar
file containing informant-core and a set of plugins.  The 'packaged'
plugins should be displayed separately in plugins.html, since they
cannot be uninstalled.

Create a maven plugin to make it super easy to create a package with a
given set of informant plugins.