Expose username and session attribute properties
Use the new plugin properties mechanism in informant-core to expose
these properties.

'sessionUsernameAttribute' for specifying a (possibly nested) path to
the username.

'sessionAttributes' for specifying (possibly nested) paths to session
state to expose in the root context map. Allows a special value of '*'
to expose all top-level session attributes (calling toString() on each).