Replace statement mirror map with aspectj mixin
Statement mirror map is showing up as a bottleneck when reading millions
of records. Replace this with an aspectj mixin (inter-type). The aspectj
mixin approach is actually simpler, but was not taken originally in
order to avoid having to weave the jdbc driver when using compile-time
weaving. Only load-time weaving is supported now, so this is a much
better approach (in addition to being more performant).