Replace aspectj with direct asm bytecode weaving
This is a big change and done reluctantly. AspectJ is an awesome
library, but the bytecode weaving was adding significant overhead at
startup which is probably ok in production, but annoying in development.
Also, in apps that use temporary class loaders to run dynamic stuff,
this overhead extends beyond just startup.

Since informant's bytecode weaving needs are relatively modest, it is
not too bad to implement the bytecode weaving directly using the (also
awesome) ASM library.

This change should have continued payoff in the ability to further
customize and optimize the bytecode weaving specifically for informant
(at the continued cost of additional code maintenance).