Add properties to control fine grained metrics
The metrics for ResultSet.next() and ResultSet.get*() can get called
tens of millions of times in a single trace in some applications, and
the overhead of System.nanoTime() in this situation can become
noticeable. Provide properties to enable/disable these metrics.
ResultSet.next() should be on by default since it is very useful, but
ResultSet.get*() should be off by default since it is typically less
useful.