Handle locked H2 database more gracefully
Currently an exception is thrown and the JVM exits if the H2 database is
locked.

This causes an issue when monitoring tomcat and using JAVA_OPTS to
pass the -javaagent arg, because JAVA_OPTS is also passed to the java
process that shuts down tomcat (triggered from 'catalina.sh stop'), and
since the H2 database is locked by tomcat, the 'shutdown process' fails.

It seems more friendly to just log an error and disable Informant when
the H2 database is locked.