Handle missing sql text more gracefully
Currently, an error is logged during unit testing because jdbc plugin
can not capture the sql text for a prepared statement that the h2 driver
creates internally as part of a DatabaseMetaData call. This particular
case of DatabaseMetaData calls can be handled better and there is no
need for logging anything. And in general, since drivers are free to use
PreparedStatements internally, it seems better to log this message
(outside of the known DatabaseMetaData case) as only a warning instead
of as an error.