Speed up the tests under integration-test project
Currently tests in the integration test module are not run until the
integration-test phase. While this makes sense, it means these tests
cannot be run via the much faster 'mvn test' which avoids the expensive
package phase where the bytecode shading takes place (in the
informant-core module). And these are (currently at least) the best set
of tests so 'mvn test' without them is not really that useful.

Also, the integration-test project tests are run twice, once using
ClassLoader-based weaving which is a bit faster, and definitely much
nicer for debugging inside of eclipse, and once using javaagent-based
weaving which is more realistic of the true Informant usage scenario.

So it seems like a good compromise to move the first pass (using
ClassLoader-based weaving) to the test phase and to keep the second pass
(using javaagent-based weaving) at the integration-test phase.

This makes 'mvn test' (and especially the parallel 'mvn -T 1C test') a
useful and speedy alternative to running a full build.