Handle weaver edge case
During weaving the super types of the class that is being woven need to
be identified (in order to identify any potential pointcuts that need to
be woven, e.g. to identify if the class being woven implements
javax.servlet.Servlet.service()).

Ideally, the super types could just be loaded via Class.forName(), but
ClassFileTransformer.transform() is not re-entrant, so if the super type
hasn't already been loaded, then loading it during weaving will cause
the super type itself to not be woven.

So the fallback is to parse the bytecode for the super types. However,
this requires that the ClassLoader expose the bytecode for the super
types via ClassLoader.getResource(). Most class loaders do, but some
don't.

The purpose of this commit is to handle class loaders that don't expose
the bytecode via ClassLoader.getResource(). At least some (most?) of
the time.

It does this in two ways. First, the weaver caches all types that it has
seen (per class loader). Then when it needs to look up a super type, it
calls ClassLoader.findLoadedClass() which doesn't suffer the above
problem of calling Class.forName() since it will only return the class
if it has already been loaded. Then the type can be looked up in the
type cache using its class loader. In cases where the super type has
already been loaded (which seems to be quite often in practice), this
bypasses the need to lookup and parse the bytecode for the super type
which both bypasses the ClassLoader.getResource() problem, as well as
provides more performant weaving.

The second way the weaver tries to handle these troublesome class
loaders is a bit troublesome in its own way. When the super type has not
already been loaded by the class loader and
ClassLoader.findLoadedClass() return null, the weaver gambles and calls
Class.forName() to load the super type anyways. Some custom class
loaders (at least groovy's CallSiteClassLoader) delegate to another
class loader instead of subclassing and so Class.forName() may end up
returning a previously loaded class anyways, in which case everything
is good and we can proceed with no problems. If Class.forName() returns
a new class that wasn't previously loaded, then it (probably?) wasn't
woven and so a warning message is logged and the super type info is
gathered via reflection instead of via parsing the bytecode.