Convert jdbc-plugin unit tests to hsqldb
Now that IsolatedWeavingClassLoader behaves more consistently with
WeavingClassFileTransforer, informant is started up first before weaving
kicks in, which means the h2 classes are not woven. This is fine when
running against shaded informant-core, since then it is the shaded h2
classes that are not woven, but when running inside an IDE against
unshaded informant-core (which is convenient for debugging), the tests
fail since the driver is now not woven. Switching to another jdbc driver
(in this case hsqldb) resolves the issue.