Push weaving metric down into the weaver
Currently the weaving metric is captured in
WeavingClassFileTransformer.transform(), but not in
IsolatedWeavingClassLoader.

Pushing the weaving metric down into the weaver will give more
equivalent behavior for both WeavingClassFileTransformer and
IsolatedWeavingClassLoader.

Also, this will eliminate WeavingClassFileTransformer's dependency on
PluginServicesImpl, which is currently requiring many classes to be
pre-initialized due to ClassFileTransformer.transform() not being
allowed to trigger any class loading.