Fix ClassFileTransformer class pre-initialization
The existing algorithm for calculating 'used' classes does not look at
classes referenced inside of method bodies (oops). This was a simple
change, but then it exploded the number of 'used' classes and slowed
down startup time significantly.

The best solution seems to be to calculate the 'used' classes more
precisely (following method calls recursively), and then to
pre-initialize this hard-coded list of classes at startup. Since this
hard-coded list can get easily out of sync, add a unit test to
re-calculate the list of 'used' classes and compare it to the hard-coded
list.

The results of the 'used' calculation were verified by comparing them
against the mature proguard tool. It was tempting to use proguard itself
to perform this calculation in the unit test, but proguard is GPL and
prints out a big GPL warning message when it's API is used outside of
the tool itself, and I'm concerned this would scare people to see this
message when they build informant.