Replace modals with expanding blocks, fix export
Oh well, now that modals work, they need to be replaced. Expanding
sections look much better and are more consistent.

Also, embed the 'modal' data directly into the trace json instead of
fetching it via ajax. This is needed in order to fix the export page. If
needed, at some point the traces json could be trimmed on the server
when serving to the traces page by replacing the expanded content by the
rolling file offset/length of the data to be fetched via ajax. (this
could even include more sections, like the expanded span messages which
can be large as well)