Expose limited regex to Pointcuts
Currently full regex is available by surrounding typeName or methodName
with forward slashes. The problem with full regex is that . and $ are
commonly used in class names and so must be escaped ( $ is the bigger
problem). Keep full regex (for now at least) to handle jdbc plugin's
set parameter pointcut, /(?!setNull$)set.*/.