Remove PluginServicesProxy
This required an additional volatile read for all PluginServices access
and made things generally more complex.

It was needed because h2 was triggering java.sql.DriverManager which was
loading other jdbc drivers before Informant was ready to handle
jdbc-plugin. This turns out to be a recent change in h2 that is
triggering java.sql.DriverManager, so rolling back a minor version
addresses the issue in a much more low impact way.

Added unit test to ensure java.sql.DriverManager initialization doesn't
creep back.