Don't capture jdbc connection close by default
It is not very interesting, usually very fast just returning the
connection to a connection pool. Also, often connection is wrapped by
proxy which doesn't pass close on to the wrapped connection, and proxies
aren't woven as a performance optimization (see
WeavingClassFileTransformer.java), in which case "jdbc connection close"
is not captured anyways.